Splits Prototype modules into separate files
