+++ lib/openstudio-standards/prototypes/common/buildings/Prototype.LargeOffice.rb
@@ -1,0 +1,61 @@
+
+# Custom changes for the LargeOffice prototype.
+# These are changes that are inconsistent with other prototype
+# building types.
+module LargeOffice
+  def model_custom_hvac_tweaks(building_type, climate_zone, prototype_input, model)
+    system_to_space_map = define_hvac_system_map(building_type, climate_zone)
+
+    system_to_space_map.each do |system|
+      # find all zones associated with these spaces
+      thermal_zones = []
+      system['space_names'].each do |space_name|
+        space = model.getSpaceByName(space_name)
+        if space.empty?
+          OpenStudio.logFree(OpenStudio::Error, 'openstudio.model.Model', "No space called #{space_name} was found in the model")
+          return false
+        end
+        space = space.get
+        zone = space.thermalZone
+        if zone.empty?
+          OpenStudio.logFree(OpenStudio::Error, 'openstudio.model.Model', "No thermal zone was created for the space called #{space_name}")
+          return false
+        end
+        thermal_zones << zone.get
+      end
+
+      return_plenum = nil
+      unless system['return_plenum'].nil?
+        return_plenum_space = model.getSpaceByName(system['return_plenum'])
+        if return_plenum_space.empty?
+          OpenStudio.logFree(OpenStudio::Error, 'openstudio.model.Model', "No space called #{system['return_plenum']} was found in the model")
+          return false
+        end
+        return_plenum_space = return_plenum_space.get
+        return_plenum = return_plenum_space.thermalZone
+        if return_plenum.empty?
+          OpenStudio.logFree(OpenStudio::Error, 'openstudio.model.Model', "No thermal zone was created for the space called #{system['return_plenum']}")
+          return false
+        end
+        return_plenum = return_plenum.get
+      end
+    end
+
+    return true
+  end
+
+  def update_waterheater_loss_coefficient(model)
+    case template
+      when '90.1-2004', '90.1-2007', '90.1-2010', '90.1-2013', 'NECB 2011'
+        model.getWaterHeaterMixeds.sort.each do |water_heater|
+          water_heater.setOffCycleLossCoefficienttoAmbientTemperature(11.25413987)
+          water_heater.setOnCycleLossCoefficienttoAmbientTemperature(11.25413987)
+        end
+    end
+  end
+
+  def model_custom_swh_tweaks(model, building_type, climate_zone, prototype_input)
+    update_waterheater_loss_coefficient(model)
+    return true
+  end
+end
