+++ lib/openstudio-standards/prototypes/common/buildings/Prototype.PrimarySchool.rb
@@ -1,0 +1,54 @@
+
+# Custom changes for the PrimarySchool prototype.
+# These are changes that are inconsistent with other prototype
+# building types.
+module PrimarySchool
+  def model_custom_hvac_tweaks(building_type, climate_zone, prototype_input, model)
+    OpenStudio.logFree(OpenStudio::Info, 'openstudio.model.Model', 'Started building type specific adjustments')
+
+    #
+    # add extra equipment for kitchen
+    add_extra_equip_kitchen(model)
+    return true
+  end
+
+  # add extra equipment for kitchen
+  def add_extra_equip_kitchen(model)
+    kitchen_space = model.getSpaceByName('Kitchen_ZN_1_FLR_1')
+    kitchen_space = kitchen_space.get
+    kitchen_space_type = kitchen_space.spaceType.get
+    elec_equip_def1 = OpenStudio::Model::ElectricEquipmentDefinition.new(model)
+    elec_equip_def2 = OpenStudio::Model::ElectricEquipmentDefinition.new(model)
+    elec_equip_def1.setName('Kitchen Electric Equipment Definition1')
+    elec_equip_def2.setName('Kitchen Electric Equipment Definition2')
+    case template
+      when '90.1-2004', '90.1-2007', '90.1-2010', '90.1-2013'
+        elec_equip_def1.setFractionLatent(0)
+        elec_equip_def1.setFractionRadiant(0.25)
+        elec_equip_def1.setFractionLost(0)
+        elec_equip_def2.setFractionLatent(0)
+        elec_equip_def2.setFractionRadiant(0.25)
+        elec_equip_def2.setFractionLost(0)
+        if template == '90.1-2013'
+          elec_equip_def1.setDesignLevel(915)
+          elec_equip_def2.setDesignLevel(570)
+        else
+          elec_equip_def1.setDesignLevel(99_999.88)
+          elec_equip_def2.setDesignLevel(99_999.99)
+        end
+        # Create the electric equipment instance and hook it up to the space type
+        elec_equip1 = OpenStudio::Model::ElectricEquipment.new(elec_equip_def1)
+        elec_equip2 = OpenStudio::Model::ElectricEquipment.new(elec_equip_def2)
+        elec_equip1.setName('Kitchen_Reach-in-Freezer')
+        elec_equip2.setName('Kitchen_Reach-in-Refrigerator')
+        elec_equip1.setSpaceType(kitchen_space_type)
+        elec_equip2.setSpaceType(kitchen_space_type)
+        elec_equip1.setSchedule(model_add_schedule(model, 'SchoolPrimary ALWAYS_ON'))
+        elec_equip2.setSchedule(model_add_schedule(model, 'SchoolPrimary ALWAYS_ON'))
+    end
+  end
+
+  def model_custom_swh_tweaks(model, building_type, climate_zone, prototype_input)
+    return true
+  end
+end
